# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentations Notes",
        "version": "1.0.0"
    },
    "paths": {
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get notes",
                "description": "Display all data notes",
                "operationId": "15d18a5b7eb25ecf34906490be11ab4d",
                "responses": {
                    "200": {
                        "description": "Fetch notes success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create new note",
                "description": "Insert new note data",
                "operationId": "a56780182e2924912f3664ae75a288ea",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Submit notes success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/{note_id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get note by ID",
                "description": "Display specific note by ID",
                "operationId": "49a9e528dc33d46f1df569292d2ebd70",
                "parameters": [
                    {
                        "name": "note_id",
                        "in": "path",
                        "description": "ID of notes",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "da9081e0-b20b-4228-a605-4afa34b8e963"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch notes success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/update/{note_id}": {
            "patch": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update note by ID",
                "description": "Modified note data by ID",
                "operationId": "a9933ff17c2af220ef413cc07d5308d5",
                "parameters": [
                    {
                        "name": "note_id",
                        "in": "path",
                        "description": "ID of notes",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "da9081e0-b20b-4228-a605-4afa34b8e963"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update notes success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/delete/{note_id}": {
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get note by ID",
                "description": "Display specific note by ID",
                "operationId": "e167001e88080dc25136b1af498b2035",
                "parameters": [
                    {
                        "name": "note_id",
                        "in": "path",
                        "description": "ID of notes",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "da9081e0-b20b-4228-a605-4afa34b8e963"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete notes success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}